Use TargetSdkLevel in CalculateVerificationInfo
We were using api_level=0 (meaning most recent api) for calculating
the verification information of a single method dynamically. This
might in the future lead to problems if we condition more of
verification on the API level. To prevent this future problem we will
instead use TargetSdkVersion as the API level.

Test: ./test.py --host

Change-Id: Ie87e0b1b6b07b43d2dc84101cb7764cc4087aca9
Signed-off-by: DennySPb <dennyspb@gmail.com>